requirements.txt
aiohttp==3.8.0
aiosignal==1.2.0
aniso8601==9.0.1
async-timeout==4.0.0
asynctest==0.13.0
attrs==21.2.0
base58==2.1.1
bitarray==1.2.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
cytoolz==0.11.0
dataclasses==0.8
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
frozenlist==1.2.0
hexbytes==0.2.2
idna==3.3
idna-ssl==1.1.0
importlib-metadata==4.8.1
importlib-resources==5.4.0
ipfsapi==0.4.4
ipfshttpclient==0.8.0a2
itsdangerous==2.0.1
Jinja2==3.0.2
jsonschema==3.2.0
lru-dict==1.1.7
MarkupSafe==2.0.1
multiaddr==0.0.9
multidict==5.2.0
netaddr==0.8.0
numpy==1.19.5
pandas==1.1.5
parsimonious==0.8.1
Pillow==8.4.0
protobuf==3.19.1
pycryptodome==3.11.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
rlp==2.0.1
six==1.16.0
toolz==0.11.1
typing-extensions==3.10.0.2
urllib3==1.26.7
varint==1.0.2
web3==5.24.0
websockets==9.1
Werkzeug==2.0.2
yarl==1.7.2
zipp==3.6.0
